Fix postal code validation on default warehouse page
ISSUE: PLCR16-12

+++ src/BusinessLogic/Resources/js/DefaultWarehouseController.js
@@ -167,7 +167,10 @@ var Packlink = window.Packlink || {};
             }
 
             postalCodeInput.value = currentPostalCode + ' - ' + currentCity;
-            templateService.removeError(postalCodeInput);
+            if (currentPostalCode !== '' && currentCity !== '') {
+                templateService.removeError(postalCodeInput);
+            }
+
             utilityService.configureInputElements();
         }
